chore: Export correct createAnimatedComponent (#8395)
## Summary

Up to this PR classic `createAnimatedComponent` could only be "imported"
and used by:

```tsx
import Animated from 'react-native-reanimated'

const Test = Animated.createAnimatedComponent(Text)
```

And when user decided to import `createAnimatedComponent` directly it
resulted in importing `createAnimatedComponent` from CSS. It created
issues and confusion.

We've decided to also export classic `createAnimatedComponent` as a
standalone function, so it can be used directly as
`createAnimatedComponent(Text)` and also as
`Animated.createAnimatedComponent(Text)`

This also lead to renaming `createAnimatedComponent` from CSS to
`createCSSAnimatedComponent`, for better clearance.


## Test plan