feat: Add working macos-example@0.79 (#8248)
## Summary

This PR brings back to life macos-example. It's working on the latest
`react-native-macos` stable version - `0.79`. The version is patched
because of usage of depracated function that broke animations. I've
already opened a
[PR](https://github.com/microsoft/react-native-macos/issues/2705) in
`react-native-macos` repo, and when (and if) it will be resolved we
should update deps accordingly.



https://github.com/user-attachments/assets/5fa9c566-d230-463b-8a67-a45f9cafa9d0

## Test plan

Run macos-example and check some examples.