chore: use same versions of dependencies in monorepo where applicable (#7702)
## Summary

This PR aims to improve security and stability in the monorepo.

Key changes:

- Bumped all common dependencies (like eslint, prettier, typescript) to
the same exact versions. This fixes issues where one workspace would use
a different version, potentially leading to conflicts or build issues.
- All dependencies were made exact, meaning no caret^ or tilde~
matchers. This will improve security against supply-chain attacks and
resolve dependency duplication issues.
   - Added a CI step that disallows caret^ and tilde~ matchers. 
- Deduped dependencies.
- I also added a step to yarn validation CI
[here](https://github.com/software-mansion/react-native-reanimated/pull/8383)
that fails if the dependencies are dupes. With exact versions of
dependencies duplication shouldn't happen. With the use of matches the
duplication arises in the following scenario:
1. Workspace A installs a package, i.e. `foo@^1.0.0`. The available
version on NPM is `foo@1.1.1` which in installed.
2. Later on, second workspace, B, installs the same package,
`foo@^1.0.0`. Newer version on NPM, `foo@1.2.2` is available and
installed. The version used by workspace A **is not changed**.
3. Now two versions of `foo`, 1.1.1 and 1.2.2 are installed in the repo.
- This was especially painful for exotic apps like tvos or macos which
were pulling conflicting dependencies.
   - To fix the CI error simply run `yarn dedupe` and commit changes.
 
- Amended the config of Yarn:
   - Now all dependencies are by default installed as exact versions.
   - Type packages aren't installed by default.
   - package.json hooks (`postinstall` etc.) are explicitly disabled.

- Dependency cleanup:
   - All eslint plugins moved to the root of the monorepo.
   - Fixed all fixable peer-dependency problems.
- Moved common dependencies to the root of the monorepo/removed unused
dependencies.

## Test plan

CI

I ran the example apps and the docs and they run fine.