Fix synchronous props updates on iOS below RN 0.81 (#8427)
## Summary

Currently, animating props synchronously is broken on React Native 0.80
or older because `schedulerDidSynchronouslyUpdateViewOnUIThread:props:`
method that we're using was added
https://github.com/facebook/react-native/pull/51470 in which was
released in React Native 0.81.

In
https://github.com/software-mansion/react-native-reanimated/pull/8248,
@patrycjakalinska added ifdefs to disable this flow on
`react-native-macos` but it is not a problem with macOS itself but
rather with the version of React Native (currently, `macos-example` runs
on 0.79).

This PR fixes animating props synchronously on iOS on React Native 0.80
and older.

## Test plan