fix: web types (#8401)
Required for 
- #8371

## Summary

This PR is an ongoing effort to reduce the impact of Reanimated and
Worklets in web bundles and to simplify the logic of platform-dependent
code.

The ideal solution is to have native-only files with a `.native.ts`
extension and web only files with just `.ts` extension alongside them.
However, currently we also have `.web.ts` files, which can be
problematic in some environments. We want to transition to `.native.ts`
files gradually.

This PR does a few things:

- It edits TSConfigs alongside the repository. By default we specify
that accepted extensions are `.native.ts`, `.ios.ts`, `.android.ts`,
`.ts` (obviously also `.tsx` etc.). This means that when in IDE we click
`go to source` we'll go to a `.native` file which is the most common use
case for us. Also types for type-checking will be pulled like this.

- We add additional TSConfigs for web types. These extensions ignore
`.native` .etc. extensions and accept `.web` extensions. You can switch
to those TSConfigs for web type-checking in your IDE.

- Adding type checking scripts and CI checks for web types.

- Fixing (or ignoring) all current problems with web types - i.e.
mismatching web and native types in Reanimated.

- Amending eslint configs to properly work in this setup.

## Test plan

CI