chore: Remove missing animationName CSS prop warning (#8448)
## Summary

I decided to remove this one warning as it shows up for all components
that use `react-native-css` library for styling and specify e.g. the
`duration-100` class, which is converted to both: `animationDuration`
and `transitionDuration` properties.

I decided to remove the warning on our side as it is not that much
useful as remaining 2 warnings that still exist and seems to be simpler
than trying to add complex logic on the `react-native-css` (or
dependent) library side.