chore(Worklets): remove runtime-string association from RuntimeManager (#8443)
## Summary

I'm removing the association of runtime name - worklet runtime from
runtime manager. It was added as a possibility for future extension but
since the names of the runtimes don't have to be unique and there are
already unique IDs for the runtimes, the association is obsolete.

It's safe to delete it since it's not used and it's a private API.

## Test plan

CI