fix: Animated style and switch not working in Freeze Example on web (#8467)
## Summary

As the title states. It seems that the `onTouchEnd` prop on the `View`
component doesn't work with `react-native-web`. We can't also animate
the `left` property on views with `static` or `relative` position on web
(it only works for views with `absolute` position or is applied when the
view is rendered), thus I replaced it with the `transform` property that
works for all view position types.

## Test plan

| Before | After |
|-|-|
| <video
src="https://github.com/user-attachments/assets/6f91f4ce-ba86-4283-9a63-d6d53e7324d6"
/> | <video
src="https://github.com/user-attachments/assets/0fcc9981-0535-433d-b503-2e4440d82a81"
/> |

>[!note]
> You can see that the CSS animation restarts every time the screen is
entered. This happens because of renders being triggered when the screen
is mounted and new renders trigger the new animation to start. I will
experiment to make this behavior more consistent with how it works on
mobile to ensure that the animation is not restarted when the screen is
reentered.