feat(Worklets): mock implementation for Jest (#8465)
## Summary

Unfortunately a typical Jest setup in React Native pulls code from
`.native` files which don't make sense to be tested and implemented for
Jest environment. For this case we implement a full mock of Worklets to
use in testing.

Extracted from:
- #8371

In another PR all uses of `IS_JEST` will be removed from Worklets. 

## Test plan

Run Jest tests in Worklets and Reanimated.