feat: Add support for `filter` property (#8451)
## Summary

This PR adds support for `filter` property. Before it was only supported
on web.

It adds `processFilter` function, adjusts web implementation of this
function and changes the name to `processFilterWeb`. I've also added
some tests. The filter property is also added to native CSS props.


TO-DO:

- [x] add support for `dropShadow`
- [x] add tests
- [x] add support for CSS #8483 
- [x] fix animate filter example
- [x] check `dropShadow` as an object
- [x] check web and fabric
- [x] check animations

> [!IMPORTANT]
>**Disclaimer** sometimes there is a problem with precision as the
outcome is for example `3.237473234e-14px` and it is non parsable -
maybe we can change global precision to some defined number of decimal
points, or locally I can change it inside filter function.

> [!IMPORTANT]
>**Disclaimer 2** sometimes we the static asset does not work (most of
the time) and it's not related to the `filter` processing, when testing
please replace `balloonsImage` with some working `uri`. The asset
problem will be fixed in some other PR.

## Test plan

1. Build app across platforms. 
2. Test `Animate Filter` example in Reanimated App on Android, iOS and
Web.
4. Run tests.