chore: Make obtaining logger config safer (#8527)
## Summary

This PR adds a helper `getLoggerConfig` function that returns the
current logger config stored under `global.__reanimatedLoggerConfig` if
it exists or stores the default config in the
`global.__reanimatedLoggerConfig` variable before returning it. This
approach is much safer than the previous one which assumed that the
`registerLoggerConfig` is executed on the UI runtime before the logger
is used.

I decided to rewrite this logic as the issue I tried to fix in the #8227
PR seems to still persist (see
[this](https://github.com/expo/expo/issues/39582#issuecomment-3481631086)
comment).