fix: Use correct color processor in CSS animations and transitions (#8566)
## Summary

The incorrect color processor was used before, which broke the
interpolation to/from the `transparent` color. Instead of returning the
`'transparent'` string for `transparent` colors, the `undefined` value
was returned, which was treated as a missing value in CSS transitions,
thus the color was interpolated to the default fallback value, which is
`black`.