feat: Add support for filter property to CSS (#8483)
## Summary

This PR adds support for `filter` property to CSS. 

It integrates `processFilter` function into CSS and prepares native side
for `filter` property.


TO-DO:

- [x] add support for CSS
- [x] check web and fabric

> [!IMPORTANT]
>**Disclaimer** sometimes there is a problem with precision as the
outcome is for example `3.237473234e-14px` and it is non parsable -
maybe we can change global precision to some defined number of decimal
points, or locally I can change it inside filter function.

> [!IMPORTANT]
>**Disclaimer 2** sometimes we the static asset does not work (most of
the time) and it's not related to the `filter` processing, when testing
please replace `balloonsImage` with some working `uri`. The asset
problem will be fixed in some other PR.

## Test plan

1. Build app across platforms. 
2. Test `Filter` example in CSSApp on Android, iOS and Web.
4. Run tests.

---------

Co-authored-by: Mateusz Łopaciński <lop.mateusz.2001@gmail.com>