Add detached animated styles example (#8577)
## Summary

This PR adds "Detached animated styles" example which demonstrates the
usage of two animated styles that are attached or detached depending on
the React state. Currently, the implementation is partially broken, as
demonstrated on the video recording below:

1. Animating `backgroundColor` works only from the first attached
animated style. When animated style that drives `width` is animated
first, then animating `backgroundColor` doesn't work. It only "jumps" to
`lime` or `blue` once the animation has finished, without displaying the
intermediate colors.
2. The animated `width` and `backgroundColor` properties dont't return
to the original value (`100` and `black`, respectively) when detached.


https://github.com/user-attachments/assets/4365574f-4dd8-4db6-a984-6651696f2005
## Test plan

---------

Co-authored-by: Mateusz Łopaciński <lop.mateusz.2001@gmail.com>