chore: Pass labelTypes to CSS example screens via props (#8588)
## Summary

In the previous implementation, the same array of `labelTypes` was
duplicated between the screen implementation and the route configuration
object. After these changes, we can access the `labelTypes` prop
specified in the route configuration object via the screen props. This
reduces duplication and makes the implementation safer as the same
labels as specified in the routes will be used in the screen.