fix: Incorrect transparent color behavior in CSS transitions (#8542)
## Summary

This PR fixes the incorrect transparent color interpolation issue in CSS
transitions introduced in the #8433 PR, which replaced the
`'transparent'` color string with the `null` value. The issue stems from
the fact that we treat the `null` value in the `folly::dynamic` type
used to store the view's tyle in cpp as the empty value, so when the
color was set to `transparent`, the value was considered as non-existent
in CPP and the black color was used as a fallback (as this is the
default border color).

Since I don't want to use the `'transparent'` string to distinguish the
`transparent` color from any other color, and no numeric value can be
used as all values correspond to different valid colors, I came up with
an idea to use the boolean `false` value. Since a boolean `false` value
is essentially the same as `0`, it can be safely used in worklet-based
animations, where the value will be treated as `0` (black transparent)
after committing to the ShadowTree. At the same time, we can distinguish
it from other numeric colors in CSS animations/transitions by checking
if the value is a boolean.

This PR also restores the use of the `null` value for invalid colors,
which was changed from `null` to `undefined` in the #8433 PR, in which I
wanted to represent the `transparent` color as `null`. Since the
`transparent` color is now represented by the `false` value, I could
restore these changes.

## Example recordings

These recordings show the issue with CSS transitions that this PR fixes
- look at the **In Circle** example.

| Before | After |
|-|-|
| <video
src="https://github.com/user-attachments/assets/1c7b8a2e-465f-4105-937c-d09a5420a4ac"
/> | <video
src="https://github.com/user-attachments/assets/d0aa888b-cebd-4296-9768-e280d0410bbd"
/> |

## Test plan

I went through all color-related examples in the Reanimated tab in the
example app and in the CSS example app.

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>