fix(Worklets): strip RN Runtime assertions only for other runtimes (#8531)
## Summary

After the changes done in #8471 it turned out that in Bundle Mode there
were redundant checks for cpp version on Worklet Runtimes and also
freezing Serializables on Worklet Runtimes. Check for the cpp version
started to crash the runtime (previously it was silently failing) while
the freezing could in some circumstances break the logic of Mutables.

## Test plan

Run the app in Bundle Mode and see it's no longer crashing.