chore: Move common parts of transforms and filters style interpolator to the base class (#8590)
## Summary

This PR moves duplicate code from the filter and transform related
interpolators to the common base classes.

We decided with @patrycjakalinska to copy most of the logic from the
transform operations interpolator implementation while implementing the
filter property interpolator in #8483 and to clean up the code in a
follow-up PR, as the PR with the filter interpolator implementation was
complex enough and we didn't want to introduce more complexity.