Configure OIDC in nightly publish action (#8653)
## Summary

This PR configures nightly publish action for reanimated to use OIDC
authentication method.

This is based on a similar change in react-native-screens:
https://github.com/software-mansion/react-native-screens/pull/3220

## Test plan

There's no good way to test this. After merging will run publish action
(or wait till nightly) to see if it succeeds. In the worst case we won't
publish one nightly release.