fix(Reanimated): false positive animated style in non-animated component error in InlinePropManager (#8677)
## Summary

#8662 added a QoL error when passing an animated style to non-animated
component but this triggers a false positive in `InlinePropsManager`
because it goes over `Object.entries` of every style object in the
component. I just simply omit animated styles before the throwing
property is accessed.

## Test plan

I went over several examples in the Example App on native and web and
all was gucci 👌