Shared Element Transition on the New Architecture (feature flag) (#7466)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR brings back the Shared Element Transition feature to reanimated
(within a limited scope and behind a feature flag).

Current limitations:
- animations with native modals don't work on iOS (the modal is
displayed over the Shared Element). This can be fixed by plugging the
Shared Element into the Window, instead of the React's RootVIew.
- can't define custom progress (swipe back) transitions (it just isn't
implemented yet, it was easier to hard code)
- progress transitions animate less props than basic ones (for the same
reason as above)
- on iOS there are still some issues with header height (e.g. the `[SET]
Modals`) example
- there are some performance bottlenecks, that stem from transforms
being recalculated too eagerly (this requires some changes to how we
recognize whether to start a transition)
- we have to use a workaround for rnscreens to not keep a snapshot of
the transitioning view on the popped screen. Currently the
[fix](https://github.com/software-mansion/react-native-screens/pull/3296)
is available in their nightly release.
- zIndex of exiting animations changes (they are brought to the front).
This differs from the current behaviour.
- sometimes on a fast refresh, the transition will run for elements that
are on the same screen. From our testing it happened when the navigator
was wrapped with a context provider, but we didn't investigate it
further. The problem was, that in this case the native view that
represents the screen actually changed, so the SET logic worked as it
should. But it definitely is not the expected outcome

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>