feat(Worklets): run Bundle Mode without patches to React Native (#8744)
## Summary

Since it doesn't seem that no-copy sharing of the JS bundle is going to
be available in React Native any time soon, we have to opt into a
single-copy bundle.

Worklets are going to obtain the Bundle the same way React Native does,
either via Metro, a file or the asset manager.

## Test plan

- in Bundle Mode
   - [x] iOS Debug works
   - [x] iOS Release works
   - [x] Android Debug works
   - [x] Android Release (Assets) works
   - [x] Android Release (File) works
- in Legacy Mode
   - [x] iOS works
   - [x] Android works
   - [x] compatibility CI passes