docs: remove reading from shared value on JS thread in "your first animation" example (#8788)
## Summary

While going through the docs, I've noticed that the example of an
animation we give in the docs in the "Your First Animation" page
contains a piece of code that reads from the shared value on the JS
thread. That is something we discourage doing in some other places in
the docs (like the **Performance** page)

While discussing it with @tomekzaw the conclusion we made was that
perhaps replacing this with some variant of `Math.random` is the best
way out of this. "Your First Animation" is still supposed to be
presenting a simple case of an animation so bringing into it more APIs
(to avoid reading from shared value on JS thread) could be confusing for
new users, while `Math.random` is still keeping it simple.

For the initial width 100 in the example I've added a random width of
range 50-150 for the consecutive values.

## Test plan

Signed-off-by: Oskar Pawica <pawicao@gmail.com>